+++ packages/yourloops/models/shoreline.ts
@@ -30,7 +30,7 @@ import { Units } from "./generic";
 import { LanguageCodes } from "./locales";
 import { MedicalData } from "./device-data";
 import { HcpProfession } from "./hcp-profession";
-import { IAlert } from "./alert";
+import { Alert } from "./alert";
 
 enum UserRoles {
   hcp = "hcp",
@@ -96,7 +96,7 @@ interface IUser {
   preferences?: Preferences | null;
   /** Patient medical data. undefined means not fetched, null if the fetch failed */
   medicalData?: MedicalData | null;
-  alerts?: IAlert | null;
+  alerts?: Alert | null;
 }
 
 export { IUser, Profile, Settings, Preferences, Consent, UserRoles };