+++ packages/yourloops/test/pages/hcp/patients/patient-utils.test.tsx
@@ -25,14 +25,302 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { createPatient } from "../../../common/utils";
+import { createAlert, createPatient, createTeamMember, createTeamUser } from "../../../common/utils";
+import {
+  comparePatients,
+  mapMembersToPatientTeamStatus,
+  mapTeamUserToPatient,
+} from "../../../../pages/hcp/patients/utils";
+import { PatientTableSortFields, UserInvitationStatus } from "../../../../models/generic";
+import { Patient, PatientTeam } from "../../../../models/patient";
+import { INotification } from "../../../../lib/notifications";
+import { Profile } from "../../../../models/shoreline";
 
 describe("Patient utils", () => {
 
   describe("comparePatients", () => {
 
-    it("should return correct order when sorting patients on alertTimeTarget", () => {
-      const patient1 = createPatient("fakePatientId", []);
+    describe("alertTimeTarget", () => {
+      const smallerAlert = createAlert(10, 0);
+      const biggerAlert = createAlert(11, 0);
+
+      it("should return -1 when first patient has a smaller alert value", () => {
+        const patient1 = createPatient("fakePatient1Id", [], smallerAlert);
+        const patient2 = createPatient("fakePatient2Id", [], biggerAlert);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertTimeTarget);
+        expect(res).toBe(-1);
+      });
+
+      it("should return -1 when second patient has a smaller alert value", () => {
+        const patient1 = createPatient("fakePatient1Id", [], biggerAlert);
+        const patient2 = createPatient("fakePatient2Id", [], smallerAlert);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertTimeTarget);
+        expect(res).toBe(1);
+      });
+
+      it("should return 0 when patients have same alert value", () => {
+        const patient1 = createPatient("fakePatient1Id", [], smallerAlert);
+        const patient2 = createPatient("fakePatient2Id", [], smallerAlert);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertTimeTarget);
+        expect(res).toBe(0);
+      });
+
+      it("should return 1 when first patient has no alert", () => {
+        const patient1 = createPatient("fakePatient1Id", []);
+        const patient2 = createPatient("fakePatient2Id", [], smallerAlert);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertTimeTarget);
+        expect(res).toBe(1);
+      });
+
+      it("should return -1 when second patient has no alert", () => {
+        const patient1 = createPatient("fakePatient1Id", [], smallerAlert);
+        const patient2 = createPatient("fakePatient2Id", []);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertTimeTarget);
+        expect(res).toBe(-1);
+      });
+
+      it("should return 0 when both patient have no alert", () => {
+        const patient1 = createPatient("fakePatient1Id", []);
+        const patient2 = createPatient("fakePatient2Id", []);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertTimeTarget);
+        expect(res).toBe(0);
+      });
+    });
+
+    describe("alertHypoglycemic", () => {
+      const smallerAlert = createAlert(0, 10);
+      const biggerAlert = createAlert(0, 11);
+
+      it("should return -1 when first patient has a smaller alert", () => {
+        const patient1 = createPatient("fakePatient1Id", [], smallerAlert);
+        const patient2 = createPatient("fakePatient2Id", [], biggerAlert);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertHypoglycemic);
+        expect(res).toBe(-1);
+      });
+
+      it("should return -1 when second patient has a smaller alert value", () => {
+        const patient1 = createPatient("fakePatient1Id", [], biggerAlert);
+        const patient2 = createPatient("fakePatient2Id", [], smallerAlert);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertHypoglycemic);
+        expect(res).toBe(1);
+      });
+
+      it("should return 0 when patients have same alert value", () => {
+        const patient1 = createPatient("fakePatient1Id", [], smallerAlert);
+        const patient2 = createPatient("fakePatient2Id", [], smallerAlert);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertHypoglycemic);
+        expect(res).toBe(0);
+      });
+
+      it("should return 1 when first patient has no alert", () => {
+        const patient1 = createPatient("fakePatient1Id", []);
+        const patient2 = createPatient("fakePatient2Id", [], smallerAlert);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertHypoglycemic);
+        expect(res).toBe(1);
+      });
+
+      it("should return -1 when second patient has no alert", () => {
+        const patient1 = createPatient("fakePatient1Id", [], smallerAlert);
+        const patient2 = createPatient("fakePatient2Id", []);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertHypoglycemic);
+        expect(res).toBe(-1);
+      });
+
+      it("should return 0 when both patient have no alert", () => {
+        const patient1 = createPatient("fakePatient1Id", []);
+        const patient2 = createPatient("fakePatient2Id", []);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.alertHypoglycemic);
+        expect(res).toBe(0);
+      });
+    });
+
+    describe("remoteMonitoring", () => {
+      const smallerDate = new Date();
+      const biggerDate = new Date(smallerDate.getUTCFullYear(), smallerDate.getMonth() + 1);
+
+      it("should return -1 when first patient has a more recent date", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", smallerDate);
+        const patient2 = createPatient("fakePatient2Id", [], null, "", biggerDate);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.remoteMonitoring);
+        expect(res).toBeLessThanOrEqual(-1);
+      });
+
+      it("should return -1 when second patient has a more recent date", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", biggerDate);
+        const patient2 = createPatient("fakePatient2Id", [], null, "", smallerDate);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.remoteMonitoring);
+        expect(res).toBeGreaterThanOrEqual(1);
+      });
+
+      it("should return 0 when patients have same date", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", smallerDate);
+        const patient2 = createPatient("fakePatient2Id", [], null, "", smallerDate);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.remoteMonitoring);
+        expect(res).toBe(0);
+      });
+
+      it("should return 1 when first patient has no remote monitoring", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "");
+        const patient2 = createPatient("fakePatient2Id", [], null, "", smallerDate);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.remoteMonitoring);
+        expect(res).toBeGreaterThanOrEqual(1);
+      });
+
+      it("should return -1 when second patient has no remote monitoring", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", smallerDate);
+        const patient2 = createPatient("fakePatient2Id", [], null, "");
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.remoteMonitoring);
+        expect(res).toBeLessThanOrEqual(-1);
+      });
+
+      it("should return 0 when both patient have no remote monitoring", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "");
+        const patient2 = createPatient("fakePatient2Id", [], null, "");
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.remoteMonitoring);
+        expect(res).toBe(0);
+      });
+    });
+
+    describe("system", () => {
+      const smallerSystemName = "DBLG1";
+      const biggerSystemName = "DBLG2";
+
+      it("should return -1 when first patient has a smaller system name", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", null, smallerSystemName);
+        const patient2 = createPatient("fakePatient2Id", [], null, "", null, biggerSystemName);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.system);
+        expect(res).toBe(-1);
+      });
+
+      it("should return -1 when second patient has a smaller system name", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", null, biggerSystemName);
+        const patient2 = createPatient("fakePatient2Id", [], null, "", null, smallerSystemName);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.system);
+        expect(res).toBe(1);
+      });
+
+      it("should return 0 when patients have same system name", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", null, smallerSystemName);
+        const patient2 = createPatient("fakePatient2Id", [], null, "", null, smallerSystemName);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.system);
+        expect(res).toBe(0);
+      });
+
+      it("should return 1 when first patient has no system name", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "" );
+        const patient2 = createPatient("fakePatient2Id", [], null, "", null, smallerSystemName);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.system);
+        expect(res).toBe(1);
+      });
+
+      it("should return -1 when second patient has no system name", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", null, smallerSystemName);
+        const patient2 = createPatient("fakePatient2Id", [], null, "");
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.system);
+        expect(res).toBe(-1);
+      });
+
+      it("should return 0 when both patient have no system name", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "");
+        const patient2 = createPatient("fakePatient2Id", [], null, "");
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.system);
+        expect(res).toBe(0);
+      });
+    });
+
+    describe("patientFullName", () => {
+      const smallerPatientName = "aaron";
+      const biggerPatientName = "zacchaeus";
+
+      it("should return -1 when first patient has a smaller fullname", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, smallerPatientName);
+        const patient2 = createPatient("fakePatient2Id", [], null, biggerPatientName);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.patientFullName);
+        expect(res).toBe(-1);
+      });
+
+      it("should return -1 when second patient has a smaller fullname", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, biggerPatientName);
+        const patient2 = createPatient("fakePatient2Id", [], null, smallerPatientName);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.patientFullName);
+        expect(res).toBe(1);
+      });
+
+      it("should return 0 when patients have same fullname", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, smallerPatientName);
+        const patient2 = createPatient("fakePatient2Id", [], null, smallerPatientName);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.patientFullName);
+        expect(res).toBe(0);
+      });
+    });
+
+    describe("flag", () => {
+
+      it("should return -1 when first patient is flagged", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", null, null, true);
+        const patient2 = createPatient("fakePatient2Id", []);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.patientFullName);
+        expect(res).toBe(-1);
+      });
+
+      it("should return -1 when second patient is flagged", () => {
+        const patient1 = createPatient("fakePatient1Id", [], );
+        const patient2 = createPatient("fakePatient2Id", [], null, "", null, null, true);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.patientFullName);
+        expect(res).toBe(1);
+      });
+
+      it("should return 0 when patients are flagged", () => {
+        const patient1 = createPatient("fakePatient1Id", [], null, "", null, null, true);
+        const patient2 = createPatient("fakePatient2Id", [], null, "", null, null, true);
+        const res = comparePatients(patient1, patient2, PatientTableSortFields.patientFullName);
+        expect(res).toBe(0);
+      });
+    });
+  });
+
+  describe("mapMembersToPatientTeamStatus", () => {
+
+    it("should map correctly", () => {
+      const member = createTeamMember("fakeTeamMember", "teamName", "fakeTeamCode", UserInvitationStatus.accepted);
+      member.invitation = { id : "invitationId" } as INotification;
+      const patientTeam : PatientTeam = {
+        code : member.team.code,
+        invitation: member.invitation,
+        status: member.status,
+        teamId: member.team.id,
+        teamName: member.team.name,
+      };
+      const res = mapMembersToPatientTeamStatus(member);
+      expect(res).toStrictEqual(patientTeam);
+    });
+  });
+
+  describe("mapTeamUserToPatient", () => {
+
+    it("should map correctly", () => {
+      const profile : Profile = {
+        fullName : "fake full name",
+        firstName: "fake full",
+        lastName: "name",
+      };
+      const member = createTeamMember("fakeTeamMember", "teamName", "fakeTeamCode", UserInvitationStatus.accepted);
+      const teamUser = createTeamUser("fakeTeamMember", [member], profile);
+      const patient : Patient = {
+        alerts: null,
+        firstName: profile.firstName,
+        flagged: undefined,
+        fullName: profile.fullName,
+        lastName: profile.lastName,
+        medicalData: null,
+        remoteMonitoring: undefined,
+        system: undefined,
+        teams: [mapMembersToPatientTeamStatus(teamUser.members[0])],
+        userid: teamUser.userid,
+        username: teamUser.username,
+      };
+      const res = mapTeamUserToPatient(teamUser);
+      expect(res).toStrictEqual(patient);
     });
   });
 });